% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rboot.R
\name{boot_JB}
\alias{boot_JB}
\title{Bootstrap procedure for JB-test}
\usage{
boot_JB(x, n.boot = 1000, n.cores = 1, fix_beta = FALSE)
}
\arguments{
\item{x}{VAR object of class '\code{varx}' or any other 
that will be \link[=as.varx]{coerced} to '\code{varx}'.}

\item{n.boot}{Integer. Number of bootstrap iterations.}

\item{n.cores}{Integer. Number of allocated processor cores.}

\item{fix_beta}{Logical. If \code{TRUE}, the cointegrating vectors \eqn{\beta} 
are fixed over all bootstrap iterations. Kilian and Demiroglu (2000:43) 
suggest this for VECM with known \eqn{\beta}. 
Ignored in case of rank-unrestricted VAR.}
}
\value{
An array of dimension (\eqn{3 \times (1+K) \times} \code{n.boot}) 
  containing the \code{n.boot} iteration results for 
  \emph{(i)} the JB, skewness and kurtosis test and for
  \emph{(ii)} the multivariate and each univariate test 
  for the \eqn{K} residual time series.
}
\description{
Bootstraps the distribution of the Jarque-Bera test 
  for individual VAR and VECM as described by Kilian, Demiroglu (2000).
}
\examples{
\dontrun{
library("vars")
data("Canada")
exogen  = cbind(qtrend=(1:nrow(Canada))^2)  # quadratic trend
R.vars  = VAR(Canada, p=2, type="both", exogen=exogen)
R.resid = resid(R.vars)

# bootstrapped JB-test: p-values #
S.jb = boot_JB(x=R.vars, n.boot=5000, n.cores=1)
pvars:::test.normality(u=R.resid, distribution=S.jb[ ,"JB.stats_mlt", ])
pvars:::test.normality(u=R.resid, distribution="theoretical")
apply(R.resid, MARGIN=2, FUN=function(e_k) 
  pvars:::test.normality(u=e_k, distribution="theoretical"))
sapply(colnames(R.resid), FUN=function(k) 
  pvars:::test.normality(u=R.resid[ , k], distribution=S.jb[ , k, ]))

# density plot #
library("ggplot2")
R.data = data.frame(t(S.jb[ , "JB.stats_mlt", ]))
R.args = list(df=2*R.vars$K)
F.density = ggplot() +
  stat_density(data=R.data, aes(x=JB, color="bootstrap"), geom="line") +
  stat_function(fun=dchisq, args=R.args, n=500, aes(color="theoretical")) +
  labs(x="JB statistic", y="Density", color="Distribution", title=NULL) +
  theme_bw()
plot(F.density)
}

}
\references{
Kilian, L. and Demiroglu, U. (2000): 
  "Residual-Based Tests for Normality in Autoregressions: 
  Asymptotic Theory and Simulation Evidence", 
  \emph{Journal of Business and Economic Statistics}, 32, pp. 40-50.
}
