% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvars.R
\docType{data}
\name{EU_w}
\alias{EU_w}
\title{Weights for the \emph{Euro Monetary Policy Transmission}}
\format{
A numeric vector containing 14 named elements.
}
\source{
The prepared \emph{Eurostat} data set is directly obtainable from the 
  \emph{ZBW Journal Data Archive}: 
  \url{https://journaldata.zbw.eu/dataset/statistical-identification-in-panel-structural-vector-autoregressive-models-replication-data}
}
\usage{
data("EU_w")
}
\description{
The data set \code{EU_w} is a vector of 14 elements.
  These are weights for \eqn{N=14} member countries of the Euro area, 
  constructed as the average share of their respective real GDP 
  over the sample period in Herwartz, Wang (2024).
}
\references{
Herwartz, H., and Wang, S. (2024): 
  "Statistical Identification in Panel Structural Vector Autoregressive 
  Models based on Independence Criteria", 
  \emph{Journal of Applied Econometrics}, 39 (4), pp. 620-639.
}
\seealso{
Other data sets: 
\code{\link{ERPT}},
\code{\link{EURO}},
\code{\link{MDEM}},
\code{\link{MERM}},
\code{\link{PCAP}},
\code{\link{PCIT}}
}
\concept{data sets}
\keyword{datasets}
