% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svarirf.R
\name{irf.varx}
\alias{irf.varx}
\title{Impulse Response Functions}
\usage{
\method{irf}{varx}(x, ..., n.ahead = 20, normf = NULL)
}
\arguments{
\item{x}{VAR object of class '\code{varx}', '\code{id}', or any other 
that will be \link[=as.varx]{coerced} to '\code{varx}'.}

\item{...}{Currently not used.}

\item{n.ahead}{Integer. Number of periods to consider after the initial impulse, i.e. the horizon of the IRF.}

\item{normf}{Function. A given function that normalizes the \eqn{K x S} input-matrix 
into an output matrix of same dimension. See the example in '\link{id.iv}' 
for the normalization of Jentsch and Lunsford (2021) 
that fixes the size of the impact response.}
}
\value{
A list of class '\code{svarirf}' holding the impulse response functions as a '\code{data.frame}'.
}
\description{
Calculates impulse response functions.
}
\references{
Luetkepohl, H. (2005): 
  \emph{New Introduction to Multiple Time Series Analysis}, 
  Springer, 2nd ed.

Jentsch, C., and Lunsford, K. G. (2021):
  "Asymptotically Valid Bootstrap Inference for Proxy SVARs",
  \emph{Journal of Business and Economic Statistics}, 40, pp. 1876-1891.
}
