% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvars.R
\docType{package}
\name{pvars}
\alias{pvars-package}
\alias{pvars}
\title{pvars: VAR Modeling for Heterogeneous Panels}
\description{
This package implements (1) panel cointegration rank tests, (2) estimators for panels of VAR, and 
(3) panel-based identification methods for structural vector autoregressive (SVAR) models. 
The implemented functions allow to account for cross-sectional dependence in the residuals 
and for structural breaks in the mean and linear trend of the VAR process.
}
\details{
\bold{(1)} The panel functions to determine the cointegration rank are:
\itemize{
\item \tabular{ll}{\code{\link{pcoint.JO}}:   \tab panel Johansen procedures,}
\item \tabular{ll}{\code{\link{pcoint.BR}}:   \tab panel test with pooled two-step estimation,}
\item \tabular{ll}{\code{\link{pcoint.SL}}:   \tab panel SL-procedures,}
\item \tabular{ll}{\code{\link{pcoint.CAIN}}: \tab correlation-augmented inverse normal test.}
}\cr
\bold{(2)} The panel functions to estimate the VAR models are:
\itemize{
\item \tabular{ll}{\code{\link{pvarx.VAR}}: \tab mean-group of a panel of VAR models,}
\item \tabular{ll}{\code{\link{pvarx.VEC}}: \tab pooled cointegrating vectors in a panel VECM.}
}\cr
\bold{(3)} The panel functions to retrieve structural impact matrices are:
\itemize{
\item \tabular{ll}{\code{\link{pid.chol}}: \tab identification of panel SVAR models using Cholesky decomposition to impose recursive causality,}
\item \tabular{ll}{\code{\link{pid.grt}}:  \tab identification of panel SVEC models,}
\item \tabular{ll}{\code{\link{pid.iv}}:   \tab identification of panel SVAR models by means of proxy variables.}
\item \tabular{ll}{\code{\link{pid.dc}}:   \tab independence-based identification of panel SVAR models using distance covariance (DC) statistic,}
\item \tabular{ll}{\code{\link{pid.cvm}}:  \tab independence-based identification of panel SVAR models using Cramer-von Mises (CVM) distance.}
}\cr
Supporting tools like the panel block bootstrap procedure (\code{\link{sboot.pmb}}) 
and the provided data sets allow for the replication of the implemented literature.
}
\author{
Lennart Empting \email{lennart.empting@uni-goettingen.de}
}
