% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvars.R
\docType{package}
\name{pvars}
\alias{pvars-package}
\alias{pvars}
\title{pvars: VAR Modeling for Heterogeneous Panels}
\description{
This package implements (1) panel cointegration rank tests, (2) estimators for panels of VAR, and 
(3) panel-based identification methods for structural vector autoregressive (SVAR) models. 
The implemented functions allow to account for cross-sectional dependence in the error terms 
and for structural breaks in the deterministic terms of the VAR processes.
}
\details{
\bold{(1)} The panel functions to determine the cointegration rank are:
\itemize{
\item \tabular{ll}{\code{\link{pcoint.JO}}   \tab panel Johansen procedures,}
\item \tabular{ll}{\code{\link{pcoint.BR}}   \tab panel test with pooled two-step estimation,}
\item \tabular{ll}{\code{\link{pcoint.SL}}   \tab panel Saikkonen-Luetkepohl procedures,}
\item \tabular{ll}{\code{\link{pcoint.CAIN}} \tab correlation-augmented inverse normal test.}
}\cr
\bold{(2)} The panel functions to estimate the VAR models are:
\itemize{
\item \tabular{ll}{\code{\link{pvarx.VAR}} \tab mean-group of a panel of VAR models,}
\item \tabular{ll}{\code{\link{pvarx.VEC}} \tab pooled cointegrating vectors in a panel VECM.}
}\cr
\bold{(3)} The panel functions to retrieve structural impact matrices are:
\itemize{
\item \tabular{ll}{\code{\link{pid.chol}} \tab identification of panel SVAR models using Cholesky decomposition to impose recursive causality,}
\item \tabular{ll}{\code{\link{pid.grt}}  \tab identification of panel SVEC models,}
\item \tabular{ll}{\code{\link{pid.iv}}   \tab identification of panel SVAR models by means of proxy variables.}
\item \tabular{ll}{\code{\link{pid.dc}}   \tab independence-based identification of panel SVAR models using distance covariance (DC) statistic,}
\item \tabular{ll}{\code{\link{pid.cvm}}  \tab independence-based identification of panel SVAR models using Cramer-von Mises (CVM) distance.}
}\cr
Supporting tools, such as the specification functions (\code{\link{speci.VAR}}, 
\code{\link{speci.factors}}) and the panel block bootstrap procedure 
(\code{\link{sboot.pmb}}), complement the panel VAR functions and complete 
this coherent approach to \strong{\emph{VAR modeling for heterogeneous panels}} 
within the \strong{vars} ecosystem. The provided data sets further allow for 
the exact replication of the implemented literature.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Lenni89/pvars}
  \item Report bugs at \url{https://github.com/Lenni89/pvars/issues}
}

}
\author{
Lennart Empting \email{lennart.empting@vwl.uni-due.de}
}
